requirements.txt
playwright==1.41.2
pandas==2.2.0
loguru==0.7.2
pyyaml==6.0.1
pytest==8.0.0
openpyxl==3.1.2
python-dotenv==1.0.1
streamlit==1.32.0 